Move hooks and constants out of utils
