+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from "./useExpandedState";
+export * from "./useSearch";
+export * from "./useTableLoader";
