+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/utils.ts
@@ -1,20 +1,6 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
-import { useDeepCompareEffect, useLatest } from "react-use";
-import _ from "underscore";
-
-import {
-  skipToken,
-  useLazyListDatabaseSchemaTablesQuery,
-  useLazyListDatabaseSchemasQuery,
-  useLazyListDatabasesQuery,
-  useSearchQuery,
-} from "metabase/api";
-import { isSyncCompleted } from "metabase/lib/syncing";
-import type { IconName } from "metabase/ui";
-import type { DatabaseId, SchemaName } from "metabase-types/api";
-
 import { getUrl as getUrl_ } from "../../utils";
 
+import { CHILD_TYPES, UNNAMED_SCHEMA_NAME } from "./constants";
 import type {
   DatabaseNode,
   ExpandedState,
@@ -22,24 +8,9 @@ import type {
   ItemType,
   NodeKey,
   RootNode,
-  SchemaNode,
-  TableNode,
   TreeNode,
   TreePath,
 } from "./types";
-
-const UNNAMED_SCHEMA_NAME = "";
-const CHILD_TYPES = {
-  database: "schema",
-  schema: "table",
-  table: null,
-} as const;
-
-export const TYPE_ICONS: Record<ItemType, IconName> = {
-  table: "table2",
-  schema: "folder",
-  database: "database",
-};
 
 export function hasChildren(type: ItemType): boolean {
   return type !== "table";
@@ -55,299 +26,11 @@ export function getUrl(value: TreePath) {
   });
 }
 
-/**
- * For the currently view path, fetches the database, schema and table (or any subset that applies to the path).
- *
- * This state is managed at the top-level so we can generate a flat list of all nodes in the tree,
- * which makes virtualization possible.
- *
- * This also makes it easier to do state transformations which needs more information than is available
- * locally at any node in the path.
- *
- * This works by fetching the data and then recursively merging the results into the tree of data that was already fetched.
- */
-export function useTableLoader(path: TreePath) {
-  const [fetchDatabases, databases] = useLazyListDatabasesQuery();
-  const [fetchSchemas, schemas] = useLazyListDatabaseSchemasQuery();
-  const [fetchTables, tables] = useLazyListDatabaseSchemaTablesQuery();
-  const databasesRef = useLatest(databases);
-  const schemasRef = useLatest(schemas);
-  const tablesRef = useLatest(tables);
-
-  const [tree, setTree] = useState<TreeNode>(rootNode());
-
-  const getDatabases = useCallback(async () => {
-    const response = await fetchDatabases(
-      { include_editable_data_model: true },
-      true,
-    );
-
-    if (databasesRef.current.isError) {
-      // Do not refetch when this call failed previously.
-      // This is to prevent infinite data-loading loop as RTK query does not cache error responses.
-      return [];
-    }
-
-    return (
-      response.data?.data.map((database) =>
-        node<DatabaseNode>({
-          type: "database",
-          label: database.name,
-          value: { databaseId: database.id },
-        }),
-      ) ?? []
-    );
-  }, [fetchDatabases, databasesRef]);
-
-  const getTables = useCallback(
-    async (
-      databaseId: DatabaseId | undefined,
-      schemaName: SchemaName | undefined,
-    ) => {
-      if (databaseId === undefined || schemaName === undefined) {
-        return [];
-      }
-
-      const newArgs = {
-        id: databaseId,
-        schema: schemaName,
-        include_hidden: true,
-        include_editable_data_model: true,
-      };
-
-      if (
-        tablesRef.current.isError &&
-        _.isEqual(tablesRef.current.originalArgs, newArgs)
-      ) {
-        // Do not refetch when this call failed previously.
-        // This is to prevent infinite data-loading loop as RTK query does not cache error responses.
-        return [];
-      }
-
-      const response = await fetchTables(newArgs, true);
-
-      return (
-        response?.data?.map((table) =>
-          node<TableNode>({
-            type: "table",
-            label: table.display_name,
-            value: { databaseId, schemaName, tableId: table.id },
-            table,
-            disabled: !isSyncCompleted(table),
-          }),
-        ) ?? []
-      );
-    },
-    [fetchTables, tablesRef],
-  );
-
-  const getSchemas = useCallback(
-    async (databaseId: DatabaseId | undefined) => {
-      if (databaseId === undefined) {
-        return [];
-      }
-
-      const newArgs = {
-        id: databaseId,
-        include_hidden: true,
-        include_editable_data_model: true,
-      };
-
-      if (
-        schemasRef.current.isError &&
-        _.isEqual(schemasRef.current.originalArgs, newArgs)
-      ) {
-        // Do not refetch when this call failed previously.
-        // This is to prevent infinite data-loading loop as RTK query does not cache error responses.
-        return [];
-      }
-
-      const response = await fetchSchemas(newArgs, true);
-      return Promise.all(
-        response.data?.map(async (schemaName, _, schemas) => {
-          const schema = node<SchemaNode>({
-            type: "schema",
-            label: schemaName,
-            value: { databaseId, schemaName },
-          });
-
-          // If the schema is unnamed, or if it's the only schema in the database,
-          // fetch the tables immediately so we can render a flattened tree.
-          if (schemaName === UNNAMED_SCHEMA_NAME || schemas.length === 1) {
-            schema.children = await getTables(databaseId, schemaName);
-          }
-          return schema;
-        }) ?? [],
-      );
-    },
-    [fetchSchemas, getTables, schemasRef],
-  );
-
-  const load = useCallback(
-    async function (path: TreePath) {
-      const { databaseId, schemaName } = path;
-      const [databases, schemas, tables] = await Promise.all([
-        getDatabases(),
-        getSchemas(path.databaseId),
-        getTables(path.databaseId, path.schemaName),
-      ]);
-
-      const newTree: TreeNode = rootNode(
-        databases.map((database) => ({
-          ...database,
-          children:
-            database.value.databaseId !== databaseId
-              ? database.children
-              : schemas.map((schema) => ({
-                  ...schema,
-                  children:
-                    schema.value.schemaName !== schemaName
-                      ? schema.children
-                      : tables,
-                })),
-        })),
-      );
-      setTree((current) => {
-        const merged = merge(current, newTree);
-        return _.isEqual(current, merged) ? current : merged;
-      });
-    },
-    [getDatabases, getSchemas, getTables],
-  );
-
-  useDeepCompareEffect(() => {
-    load(path);
-  }, [
-    load,
-    path,
-    // When a table is modified, e.g. we change display_name with PUT /api/table/:id
-    // we need to manually call the lazy RTK hooks, so that the the updated table
-    // is refetched here. We detect this modification with tables.isFetching.
-    tables.isFetching,
-  ]);
-
-  return { tree, reload: load };
-}
-
-/**
- * Fetch items from the search API and renders them as a TreeNode so we can use the same
- * data structure for the tree and the search results and render them in a consistent way.
- */
-export function useSearch(query: string) {
-  const { data, isLoading } = useSearchQuery(
-    query === ""
-      ? skipToken
-      : {
-          q: query,
-          models: ["table"],
-        },
-  );
-
-  const tree = useMemo(() => {
-    const tree: TreeNode = rootNode();
-
-    data?.data.forEach((result) => {
-      const { model, database_name, database_id, table_schema, id, name } =
-        result;
-      const tableSchema = table_schema ?? "";
-
-      if (model !== "table" || database_name === null) {
-        return;
-      }
-
-      let databaseNode = tree.children.find(
-        (node) =>
-          node.type === "database" && node.value.databaseId === database_id,
-      );
-      if (!databaseNode) {
-        databaseNode = node<DatabaseNode>({
-          type: "database",
-          label: database_name,
-          value: {
-            databaseId: database_id,
-          },
-        });
-        tree.children.push(databaseNode);
-      }
-
-      let schemaNode = databaseNode.children.find((node) => {
-        return node.type === "schema" && node.value.schemaName === tableSchema;
-      });
-      if (!schemaNode) {
-        schemaNode = node<SchemaNode>({
-          type: "schema",
-          label: tableSchema,
-          value: {
-            databaseId: database_id,
-            schemaName: tableSchema,
-          },
-        });
-        databaseNode.children.push(schemaNode);
-      }
-
-      let tableNode = schemaNode.children.find(
-        (node) => node.type === "table" && node.value.tableId === id,
-      );
-      if (!tableNode) {
-        tableNode = node<TableNode>({
-          type: "table",
-          label: name,
-          value: {
-            databaseId: database_id,
-            schemaName: tableSchema,
-            tableId: id,
-          },
-          disabled: !isSyncCompleted(result),
-        });
-        schemaNode.children.push(tableNode);
-      }
-    });
-    return tree;
-  }, [data]);
-
-  return {
-    isLoading,
-    tree,
-  };
-}
-
-/**
- * Returns a state object that indicates which nodes are expanded in the tree.
- */
-export function useExpandedState(path: TreePath) {
-  const [state, setState] = useState(expandPath({}, path));
-
-  const { databaseId, schemaName, tableId } = path;
-
-  useEffect(() => {
-    // When the path changes, this means a user has navigated throught the browser back
-    // button, ensure the path is completely expanded.
-    setState((state) => expandPath(state, { databaseId, schemaName, tableId }));
-  }, [databaseId, schemaName, tableId]);
-
-  const isExpanded = useCallback(
-    (path: string | TreePath) => {
-      const key = typeof path === "string" ? path : toKey(path);
-      return Boolean(state[key]);
-    },
-    [state],
-  );
-
-  const toggle = useCallback((key: string, value?: boolean) => {
-    setState((current) => ({
-      ...current,
-      [key]: value ?? !current[key],
-    }));
-  }, []);
-
-  return {
-    isExpanded,
-    toggle,
-  };
-}
-
 // Returns a new state object with all the nodes along the path expanded.
-function expandPath(state: ExpandedState, path: TreePath): ExpandedState {
+export function expandPath(
+  state: ExpandedState,
+  path: TreePath,
+): ExpandedState {
   return {
     ...state,
     [toKey({
@@ -451,7 +134,7 @@ export function flatten(
   ];
 }
 
-function sort(nodes: TreeNode[]): TreeNode[] {
+export function sort(nodes: TreeNode[]): TreeNode[] {
   return Array.from(nodes).sort((a, b) => {
     return a.label.localeCompare(b.label);
   });
@@ -460,7 +143,10 @@ function sort(nodes: TreeNode[]): TreeNode[] {
 /**
  * Merge two TreeNodes together.
  */
-function merge(a: TreeNode | undefined, b: TreeNode | undefined): TreeNode {
+export function merge(
+  a: TreeNode | undefined,
+  b: TreeNode | undefined,
+): TreeNode {
   if (!a) {
     if (!b) {
       throw new Error("Both a and b are undefined");
@@ -492,13 +178,15 @@ function merge(a: TreeNode | undefined, b: TreeNode | undefined): TreeNode {
 /**
  * Create a unique key for a TreePath
  */
-function toKey({ databaseId, schemaName, tableId }: TreePath) {
+export function toKey({ databaseId, schemaName, tableId }: TreePath) {
   return JSON.stringify([databaseId, schemaName, tableId]);
 }
 
 type Optional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 
-function node<T extends TreeNode>(x: Optional<T, "key" | "children">): T {
+export function node<T extends TreeNode>(
+  x: Optional<T, "key" | "children">,
+): T {
   return {
     ...x,
     key: toKey(x.value),
@@ -506,7 +194,7 @@ function node<T extends TreeNode>(x: Optional<T, "key" | "children">): T {
   } as T;
 }
 
-function rootNode(children: DatabaseNode[] = []): RootNode {
+export function rootNode(children: DatabaseNode[] = []): RootNode {
   return node<RootNode>({
     type: "root",
     label: "",
@@ -515,7 +203,7 @@ function rootNode(children: DatabaseNode[] = []): RootNode {
   });
 }
 
-function loadingItem(
+export function loadingItem(
   type: ItemType,
   level: number,
   parent?: TreeNode,
